[SYCL][CUDA] Cached info from CUDA API calls to the kernel and device classes (#10104)
This PR fixes an issue with overhead when calling the same kernel
multiple times in a loop.

Right now, some calls to CUDA API happen every time the kernel is
invoked to query the same information. Those calls were moved to
`device` and `kernel` constructors and the info was cached as private
members of those classes.