[SYCL][CUDA] Pass device from context in create queue. (#10491)
Recently in the switch to UR `urQueueCreateFromNativeHandle` changed the
previous behaviour whereby a queue was created with a device taken as
the default device from the context. It changed it so that the queue was
created with the device argument instead. Since the sycl runtime always
passes a nullptr for the device when programmers call
`make_queue(nativeStream, context)`, this broke `make_queue`. This patch
reverts to the previous behaviour before the switch from pi cuda to ur
cuda.

Note that this should also fix `make_queue` for l0 which I also guess
was broken due to the asserts meaning that this line was never reached:
https://github.com/intel/llvm/blob/sycl/sycl/plugins/unified_runtime/ur/adapters/level_zero/queue.cpp#L574.
But I have not tested this.

---------

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>