[SYCL][CUDA] Do not return errors for unsupported hints
urUSMMemAdvise and urUSMPrefetch currently return
UR_RESULT_ERROR_ADAPTER_SPECIFIC when a hint is not supported.
This error code should be reserved for errors that are returned
direclty from Cuda. This commit removes all such checks from these
 entrypoints because it is common practice to just return success
when a hint is not supported (the UR specification has also been
updated to reflect that fact).